3 types of bots now exist
The code needs cleaning up and some of the functions such as highlight material need work. But all in all this is the basic function for the bots.

Moving forward you need to add different types of crops, purification of water etc but first of all the player needs a UI in order to actually collect these items from the bots so they can do something with them.