Level 1 console model finished (#4)
* console-model-initial commit

This is the initial commit for the modelling of the console for the terminal and gate room consoles.FBX

* Console model finished

This is the updated version of the console model. The model has some small personal changes to the references, as I wasn't 100% happy with how the prop designers styled some of the edges and main shapes (too rounded in places).

The model should not be UV unwrapped, but an OBJ export to FBX export may have autowrapped this correctly by happenstance. Additionally, I have UV unwarpped the screen and created a placeholder material as a small test for schematic emission materials.

* normals and additionals

There where some normal maps the wrong way around and some more details that needed adding to the model such as the back of the main screen and some rims on buttons.

Tested and passed in Unity.