Steam
Simple Steam asset added, but there is a fix to the bug

Where I changed the lighting on the Visual VFX from Transparent to Simple Lit, then Changed it to be double sided and use emission on its colour base lighting. This got rid of the green as it was ignoring the diffusion profile in any other fixes for the Transparent option I could find and still looks good and allows for the flipbook in the VFX to be used still, so its worth the trade off considering there is a limited range of hue being used in all of SGUs smoke effects and there can always be an asset flipbook 8x8 used with grater gaps in their alpha coordinates if needed.