Cabinet Devices Added to East and West Doors
Some of the archways and Hex lights have also been moved to accommodate for this, which is now a more accurate position. Therefore the only thing that needs to be done is scaling and moving the Balcony plates and re-sealing some of the gaps in the upper level walls/archways to the Butresses.