South wall and 45 South entrance walls textured
This is the the round base entrance with circles to the gate room and the 2 entrances either side of it. However this is only the walls and planes on the walls, not the doors themselves as this will be in another commit with those door being prefabs; they will all update at the same time.