Flashing lights based on Volume
The volumes weight property now controls the danger flashing lights in the level. This seems to work best since it does not alter the exposure of glowing materials and does not add unneeded 'bloom' in the same manner as exposure changes. Additionally, lights set to this script have off timings due to different units (EV LEV KELVIN)