1.3.0 (#9)
* Scorch marks added to all levels via prefab

https://trello.com/c/HOL6L7o4/133-scorchmarks-and-damage-in-level-safe-where-fires-used-to-be

* Console UI && Footsteps on Metal

* Occlusion of fires behind doors via scripting

This was applied to the Door Open script and take a list of 'culledfViaScript'. This is looped on the doors starting to open (enable fires and fire audio) and disabled when the doors are fully closed.

* Cyan issue fixed

This was due to having lift on the volume. Therefore, please avoid using lift in the future when attempting to change the weight of a volume dynamically

* WhiteOFF material change to RedOFF

* attempting to resolve conflicts